requirements.txt
# =============================================================================
# Core ML Framework
# =============================================================================
torch>=2.1.0
unsloth
transformers>=4.36.0
peft>=0.7.0
trl>=0.7.0
datasets>=2.14.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# =============================================================================
# Evaluation Metrics
# =============================================================================
rouge-score>=0.1.2
nltk>=3.8.0
evaluate>=0.4.0

# =============================================================================
# Logging & Tracking
# =============================================================================
wandb>=0.16.0

# =============================================================================
# Data Processing
# =============================================================================
pandas>=2.0.0
pyyaml>=6.0

# =============================================================================
# Visualization
# =============================================================================
matplotlib>=3.7.0
seaborn>=0.12.0

# =============================================================================
# Utilities
# =============================================================================
tqdm>=4.65.0
